move from ElasticSearch package to OpenSearch
