+++ src/Driver/DriverRegistry.php
@@ -3,7 +3,7 @@
 namespace Aternos\Model\Driver;
 
 use Aternos\Model\Driver\Cassandra\Cassandra;
-use Aternos\Model\Driver\Elasticsearch\Elasticsearch;
+use Aternos\Model\Driver\OpenSearch\OpenSearch;
 use Aternos\Model\Driver\Mysqli\Mysqli;
 use Aternos\Model\Driver\Redis\Redis;
 use Aternos\Model\Driver\Test\TestDriver;
@@ -21,7 +21,7 @@ class DriverRegistry implements DriverRegistryInterface
      */
     protected array $classes = [
         Cassandra::ID => Cassandra::class,
-        Elasticsearch::ID => Elasticsearch::class,
+        OpenSearch::ID => OpenSearch::class,
         Mysqli::ID => Mysqli::class,
         Redis::ID => Redis::class,
         TestDriver::ID => TestDriver::class
